Разбил готовый код по модулям. Сделал main.js точкой входа
