+++ js/util.js
@@ -1,0 +1,22 @@
+// Возвращает случайное целое число из переданного диапазона включительно.
+const getRandomNumber = (a, b) => {
+  const min = Math.ceil(Math.min(Math.abs(a), Math.abs(b)));
+  const max = Math.floor(Math.max(Math.abs(a), Math.abs(b)));
+  const result = Math.random() * (max - min + 1) + min;
+
+  return Math.floor(result);
+};
+
+// Возвращает случайное число с плавающей точкой из переданного диапазона включительно.
+const getRandomFloat = (a, b, digits = 1) => {
+  const lower = Math.min(Math.abs(a), Math.abs(b));
+  const upper = Math.max(Math.abs(a), Math.abs(b));
+  const result = Math.random() * (upper - lower) + lower;
+
+  return +result.toFixed(digits);
+};
+
+// Вовзращает случайный элемент массива
+const getRandomArrayElement = (elements) => elements[getRandomNumber(0, elements.length - 1)];
+
+export {getRandomNumber, getRandomFloat, getRandomArrayElement};
